@@ -1,15 +1,12 @@
 package com.example.mbaka.klicker;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.CountDownTimer;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
-import android.widget.Toast;
 
 public class PlayActivity extends AppCompatActivity {
 
@@ -34,7 +31,7 @@ public class PlayActivity extends AppCompatActivity {
 
             @Override
             public void onTick(long l) {
-                countTime.setText("Осталось: "+l/1000);
+                countTime.setText(" "+ l / 1000+" секунд");
             }
 
             @Override
